feat: #WB2-1515, create component for newPage button
